minkyung/requirements.txt
absl-py==2.1.0
addict==2.4.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bs4==0.0.1
cachetools==5.3.3
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
datasets==2.4.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
efficientnet-pytorch==0.7.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
fonttools==4.49.0
frozenlist==1.3.1
fsspec==2022.7.1
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.9.1
idna==3.3
imageio==2.34.0
imbalanced-learn==0.12.0
importlib_metadata==7.1.0
importlib_resources==6.1.3
ipykernel==6.29.3
ipython==8.12.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.22
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.6.3
jupyter_core==5.7.1
jupyter-events==0.9.1
jupyter-lsp==2.2.4
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.9.2
llvmlite==0.38.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mistune==3.0.2
multidict==6.0.2
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.2
nest-asyncio==1.6.0
networkx==3.1
notebook==7.1.1
notebook_shim==0.2.4
numba==0.55.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pooch==1.6.0
prometheus_client==0.20.0
prompt-toolkit==3.0.42
protobuf==5.26.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0
python-json-logger==2.0.7
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
randaugment==1.0.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==2.0.0
resampy==0.4.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
setuptools==59.5.0
six==1.16.0
sklearn==0.0
sniffio==1.3.1
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
stack-data==0.6.2
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.3.0
tifffile==2023.7.10
timm==0.4.5
tinycss2==1.2.1
tomli==2.0.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.4
tqdm==4.64.0
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.9
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
wget==3.2
wheel==0.41.2
widgetsnbextension==4.0.10
xxhash==3.0.0
yapf==0.40.2
yarl==1.8.1
zipp==3.17.0
